Add confirmation dialog before deleting a collection

+++ frontend/composeApp/src/commonMain/kotlin/com.earthrevealed.immaru/collections/CollectionDetailsScreen.kt
@@ -1,21 +1,34 @@
 package com.earthrevealed.immaru.collections
 
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.consumeWindowInsets
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.filled.Check
+import androidx.compose.material3.AlertDialogDefaults
+import androidx.compose.material3.BasicAlertDialog
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilledTonalButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import com.earthrevealed.immaru.common.ErrorMessage
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -24,6 +37,7 @@ fun CollectionDetailsScreen(
     viewModel: CollectionDetailsViewModel,
     onNavigateBack: () -> Unit,
 ) {
+    val showConfirmDeleteDialog = remember { mutableStateOf(false) }
     Scaffold(
         topBar = {
             TopAppBar(
@@ -68,13 +82,58 @@ fun CollectionDetailsScreen(
 
                         //TODO: Only show button if this is a persisted collection
                         FilledTonalButton(onClick = {
-                            viewModel.deleteCollection(onSuccess = { onNavigateBack() })
+                            showConfirmDeleteDialog.value = true
                         }) {
                             Text("Delete")
                         }
                     }
                 }
             }
+            if (showConfirmDeleteDialog.value) {
+                SimpleAlertDialog(
+                    "Are you sure?",
+                    onConfirm = {
+                        viewModel.deleteCollection(onSuccess = { onNavigateBack() })
+                        showConfirmDeleteDialog.value = false
+                    },
+                    onDismiss = { showConfirmDeleteDialog.value = false }
+                )
+            }
         }
     )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SimpleAlertDialog(
+    text: String,
+    onConfirm: () -> Unit,
+    onDismiss: () -> Unit,
+) {
+    BasicAlertDialog(
+        onDismissRequest = { onDismiss() },
+    ) {
+        Surface(
+            modifier = Modifier.wrapContentWidth().wrapContentHeight(),
+            shape = MaterialTheme.shapes.large,
+            tonalElevation = AlertDialogDefaults.TonalElevation
+        ) {
+            Column(modifier = Modifier.padding(16.dp)) {
+                Text(text = text)
+                Spacer(modifier = Modifier.height(24.dp))
+                Row {
+                    TextButton(
+                        onClick = { onDismiss() }
+                    ) {
+                        Text("Cancel")
+                    }
+                    TextButton(
+                        onClick = { onConfirm() },
+                    ) {
+                        Text("Confirm")
+                    }
+                }
+            }
+        }
+    }
 }
